NTBS-135: Refactor Countries seedData to seedData namespace
